Moved ‘updateAPIReady’ and ‘updateAPIStatus’ handlers from the component operator to the api operator.
