+++ source/operators/api-management/istio/createtestpod.sh
