+++ source/operators/api-management/istio/getIngressCustom.sh
