+++ source/operators/api-management/istio/manualGatewayVS.yaml
