+++ source/operators/api-management/whalecloud-apim/component-apigController-dockerfile
