--- source/operators/component-IngressController-dockerfile
@@ -1,7 +1,0 @@
-FROM python:3.7
-RUN pip install kopf
-RUN pip install kubernetes==17.14.0a1
-RUN pip install cloudevents
-ADD . /
-ADD securityController/secconkeycloak.py /
-CMD kopf run --namespace=$COMPONENT_NAMESPACE --standalone componentOperator/componentOperator.py apiOperatorSimpleIngress/apiOperatorSimpleIngress.py securityController/securityControllerKeycloak.py
