--- source/operators/component-IstioController-dockerfile
@@ -1,45 +1,0 @@
-#FROM python:3.12
-FROM python:3.12-alpine
-
-#RUN pip install kopf
-#RUN pip install kubernetes==27.2.0
-#RUN pip install cloudevents
-
-RUN pip install kopf \
-    && pip install kubernetes==27.2.0 \
-    && pip install cloudevents \
-    && pip install PyYAML \
-    && pip install requests 
-
-#Copy the componentOperator  apiOperatorIstio  securityControllerKeycloak  secconkeycloak.py code
-COPY apiOperatorIstio/apiOperatorIstio.py /apiOperatorIstio/
-COPY apiOperatorIstio/apiOperatorIstiowithApisix.py /apiOperatorIstio/
-COPY apiOperatorIstio/apiOperatorIstiowithKong.py /apiOperatorIstio/
-COPY componentOperator/componentOperator.py /componentOperator/
-COPY securityController/securityControllerKeycloak.py /securityController/
-COPY securityController/secconkeycloak.py /securityController/
-COPY apiOperatorKong/apiOperatorKong.py /apiOperatorKong/
-COPY apiOperatorApisix/apiOperatorApisix.py /apiOperatorApisix/
-COPY componentOperator/entrypoint.sh /componentOperator/
-
-RUN chmod +x componentOperator/entrypoint.sh
-
-# Setting up required ENV variables
-ARG CICD_BUILD_TIME
-ENV CICD_BUILD_TIME $CICD_BUILD_TIME
-ARG GIT_COMMIT_SHA
-ENV GIT_COMMIT_SHA $GIT_COMMIT_SHA
-
-#ADD . /
-#ADD securityController/secconkeycloak.py /
-###
-# remove --standalone option, because in combination with dependent api operator it leads to infinite loops.
-#CMD kopf run --namespace=$COMPONENT_NAMESPACE --standalone componentOperator/componentOperator.py apiOperatorIstio/apiOperatorIstio.py securityController/securityControllerKeycloak.py
-#CMD kopf run --namespace=$COMPONENT_NAMESPACE componentOperator/componentOperator.py apiOperatorIstio/apiOperatorIstio.py securityController/securityControllerKeycloak.py
-#CMD kopf run --namespace=$COMPONENT_NAMESPACE --standalone componentOperator/componentOperator.py 
-
-#This is will be used by entrypoint script to check for changesin configmap values
-RUN apk add --no-cache inotify-tools
-
-# Using the entrypoint script 
-ENTRYPOINT ["/componentOperator/entrypoint.sh"]
