+++ source/operators/component-management/.gitignore
