+++ source/operators/component-management/entrypoint.sh
