+++ source/operators/component-management/sequenceDiagrams/componentOperator.png
