+++ source/operators/component-management/test/componentOperatorTest_SMAN.py
