+++ source/operators/component-management/test/testdata_DEPAPI/CREATE_prodcat.json
