+++ source/operators/component-management/test/testdata_DEPAPI/CREATE_prodcat_2depapis.json
