+++ source/operators/component-management/test/testdata_DEPAPI/UPDATE_prodcat_2to1depapi.json
