+++ source/operators/component-management/test/testdata_DEPAPI/compop-body.json
