+++ source/operators/component-management/test/testdata_SMAN/CREATE-prodcat_sman.json
