--- source/operators/component-wso2Controller-dockerfile
@@ -1,5 +1,0 @@
-FROM python:3.7
-RUN pip install kopf
-RUN pip install kubernetes
-ADD . /
-CMD kopf run --namespace=$COMPONENT_NAMESPACE --standalone componentOperator/componentOperator.py apiOperator-wso2/apiOperator-wso2.py  securityController/securityControllerKeycloak.py
